pyproject.toml
[project]
name = "text2emoji"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "requests",
    "google-genai",
    "python-dotenv",
    "pydantic",
    "harness-wiki",
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "aiosqlite>=0.20",
    "pydantic-settings>=2.0",
]

[tool.uv.workspace]
members = ["harness-wiki"]

[tool.uv.sources]
harness-wiki = { workspace = true }
